added animated header
